Combine relayMessage into one function (#30)
99.9% of the time, messages are going to have to be relayed from a root
block now that we have the regular commit process setup. Instead of
having two separate functions for relaying a message, we should just
remove the old function that allowed relaying directly from a block. So
we don't lose any functionality, you can still relay a message directly
from a block, but you have to pass in zeros for the root block and root
block proof.