Add depositWithData to FuelERC20Gateway Contract (#31)
a new `depositWithData` function was added to help flag deposits as
going to a contract. The optional data field will have the Fuel side
fungible token contract call `process_message` with the sent coins. If
there is no data, then the coins should just be pushed to the contracts
balance with no function call.