Add missing event emit (#34)
I forgot to actually emit the event from the previous issue that added
more events to the contracts for easier analysis for the watchtower. I
also went ahead and updated the predicate root to the latest version
which isn't expected to change unless a major issue is found during
testing.